+++ exonum/src/helpers/fabric/details.rs
@@ -51,11 +51,6 @@ const PRIVATE_ALLOW_ORIGIN: &str = "PRIVATE_ALLOW_ORIGIN";
 pub struct Run;
 
 impl Run {
-    /// Returns the name of the `Run` command.
-    pub fn name() -> CommandName {
-        "run"
-    }
-
     /// Returns created database instance.
     pub fn db_helper(ctx: &Context, options: &DbOptions) -> Box<Database> {
         let path = ctx.arg::<String>(DATABASE_PATH)
@@ -117,7 +112,7 @@ impl Command for Run {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "run"
     }
 
     fn about(&self) -> &str {
@@ -158,11 +153,6 @@ impl Command for Run {
 pub struct RunDev;
 
 impl RunDev {
-    /// Returns the name of the `Run` command.
-    pub fn name() -> CommandName {
-        "run-dev"
-    }
-
     fn artifacts_directory(ctx: &Context) -> PathBuf {
         let directory = ctx.arg::<String>("ARTIFACTS_DIR")
             .unwrap_or_else(|_| ".exonum".into());
@@ -187,7 +177,7 @@ impl RunDev {
         common_config_ctx.set_arg("COMMON_CONFIG", common_config_path.clone());
         common_config_ctx.set_arg("VALIDATORS_COUNT", validators_count.into());
         let common_config_command = commands
-            .get(GenerateCommonConfig::name())
+            .get(GenerateCommonConfig.name())
             .expect("Expected GenerateCommonConfig in the commands list.");
         common_config_command.execute(commands, common_config_ctx);
 
@@ -197,7 +187,7 @@ impl RunDev {
         node_config_ctx.set_arg("SEC_CONFIG", sec_config_path.clone());
         node_config_ctx.set_arg(PEER_ADDRESS, peer_addr.into());
         let node_config_command = commands
-            .get(GenerateNodeConfig::name())
+            .get(GenerateNodeConfig.name())
             .expect("Expected GenerateNodeConfig in the commands list.");
         node_config_command.execute(commands, node_config_ctx);
 
@@ -206,7 +196,7 @@ impl RunDev {
         finalize_ctx.set_arg("SECRET_CONFIG", sec_config_path.clone());
         finalize_ctx.set_arg("OUTPUT_CONFIG_PATH", output_config_path.clone());
         let finalize_command = commands
-            .get(Finalize::name())
+            .get(Finalize.name())
             .expect("Expected Finalize in the commands list.");
         finalize_command.execute(commands, finalize_ctx);
 
@@ -237,7 +227,7 @@ impl Command for RunDev {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "run-dev"
     }
 
     fn about(&self) -> &str {
@@ -259,7 +249,7 @@ impl Command for RunDev {
 
         let new_context = exts(context);
         commands
-            .get(Run::name())
+            .get(Run.name())
             .expect("Expected Run in the commands list.")
             .execute(commands, new_context.clone())
     }
@@ -267,13 +257,6 @@ impl Command for RunDev {
 
 /// Command for the template generation.
 pub struct GenerateCommonConfig;
-
-impl GenerateCommonConfig {
-    /// Returns the name of the `GenerateCommonConfig` command.
-    pub fn name() -> CommandName {
-        "generate-template"
-    }
-}
 
 impl Command for GenerateCommonConfig {
     fn args(&self) -> Vec<Argument> {
@@ -291,7 +274,7 @@ impl Command for GenerateCommonConfig {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "generate-template"
     }
 
     fn about(&self) -> &str {
@@ -334,11 +317,6 @@ impl Command for GenerateCommonConfig {
 pub struct GenerateNodeConfig;
 
 impl GenerateNodeConfig {
-    /// Returns the name of the `GenerateNodeConfig` command.
-    pub fn name() -> CommandName {
-        "generate-config"
-    }
-
     fn addr(context: &Context) -> (SocketAddr, SocketAddr) {
         let addr_str = &context.arg::<String>(PEER_ADDRESS).unwrap_or_default();
         let error_msg = &format!("Expected an ip address in {}: {:?}", PEER_ADDRESS, addr_str);
@@ -376,7 +354,7 @@ impl Command for GenerateNodeConfig {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "generate-config"
     }
 
     fn about(&self) -> &str {
@@ -455,11 +433,6 @@ impl Command for GenerateNodeConfig {
 pub struct Finalize;
 
 impl Finalize {
-    /// Returns the name of the `Finalize` command.
-    pub fn name() -> CommandName {
-        "finalize"
-    }
-
     /// Returns `GenesisConfig` from the template.
     fn genesis_from_template(
         template: CommonConfigTemplate,
@@ -562,7 +535,7 @@ impl Command for Finalize {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "finalize"
     }
 
     fn about(&self) -> &str {
@@ -665,13 +638,6 @@ impl Command for Finalize {
 /// Command for the testnet generation.
 pub struct GenerateTestnet;
 
-impl GenerateTestnet {
-    /// Returns the name of the `GenerateTestnet` command.
-    pub fn name() -> CommandName {
-        "generate-testnet"
-    }
-}
-
 impl Command for GenerateTestnet {
     fn args(&self) -> Vec<Argument> {
         vec![
@@ -696,7 +662,7 @@ impl Command for GenerateTestnet {
     }
 
     fn name(&self) -> CommandName {
-        Self::name()
+        "generate-testnet"
     }
 
     fn about(&self) -> &str {